Fix some compiler warnings

+++ src/cunumeric/binary/binary_op_util.h
@@ -247,7 +247,10 @@ struct BinaryOp<BinaryOpCode::POWER, CODE> {
   using VAL                   = legate::legate_type_of<CODE>;
   static constexpr bool valid = true;
   BinaryOp(const std::vector<legate::Store>& args) {}
-  constexpr VAL operator()(const VAL& a, const VAL& b) const { return std::pow(a, b); }
+  constexpr VAL operator()(const VAL& a, const VAL& b) const
+  {
+    return std::pow(static_cast<double>(a), static_cast<double>(b));
+  }
 };
 
 template <>
